+++ lib/alf/errors.rb
@@ -1,0 +1,4 @@
+module Alf
+  class Error < StandardError; end
+  class NoSuchDatasetError < Error; end
+end